item: make RtfmItem a tree of GListModel
This makes RtfmItem both a GListModel and a tree of items with various
O(1) manipulation functions.

The goal (not yet achieved) is to allow the library to have a tree of
items which can be populated and manipulated by plugins. For example,
we want the platform plugin to reorder G-I information into various
platform verticals.