view: simplify widget hierarchy
If we rely on view addins for all of the view content, then we can simplify
the view widget to just be a vertical box. The addins can add their widgets
(and hide them when appropriate) such as when the language does not match
how they will want to render the data.