gir: stub out gir provider
This will be the basis of our gir plugin. The goal here is to import
data from .gir files so that we can view them in the browser.

However, I haven't decided on which of two approaches we want to use. Both
might be the answer.

1) Simply read from .gir files and build our own search indexes locally.
   This has the benefit of being quick and easy to load even personal
   projects in the documentation browser.
2) Generate a gvariant based, mmap'able file similar to what the typelib
   file is, but including extra information such as a search index. This
   would allow us to keep the RSS memory lower at the cost of VMEM when
   browsing documentation.

   The nice thing about this option is that we can add extra information
   into the data-packs on the server and allow them for download. We could
   also add a tool to deal with adding per-language overrides as custom
   metadata. This could be useful in situations where we can't easily
   parse the override signatures.

Either way, we should probably include the tools here to generate those
files from the xml .gir files.