gir: stub out rtfm-gir-compiler
The goal of this program is to parse a .gir file and generate an mmap()able
file containing a GVariant which we can read at runtime to browse docs.
The RtfmGirProvider will read from these files. How we get access to them
is still a bit up in the air. We can generate them locally or we can get
access to them via some online mechanism such as developer.gnome.org.