gir: start parsing .gir files with GMarkupParseContext
libxml2 was getting frustrating, and we don't actually need it. So instead
we can do something simpler (and much cleaner) to use GMarkupParseContext.

This is not a complete parser for .gir data yet and still has some corner
cases to be shaken out, but it lets us start playing with things.