window: add 100msec delay before search
This is meant to help coalesce queries when the user is typing fairly
quickly to help reduce load on the search index a bit.

Currently, however, the indexing layer is fast enough, but the creating/
sizing of search results is not fast enough with ListBox/ListBoxRow.

In Builder, to get around this, we had a fixed number of results. We can
do something similar in RtfmSearchResults by having a max size and ignoring
items that would fall outside the range based on scoring.

Additionally, we could do some pooling of used rows and re-use them when
we need to create a new row for the list box. Sadly, sizing will still be
pretty expensive here unless we can get to a fixed size situation with
a search result row subclass (and precalculate all our sizing info).