gir: avoid inflating search results when possible
When we reach an item in the fuzzy match that is too low of a score
to add to result set of capped size, we can avoid any further processing
of the matches (as they are sorted by score). This helps us avoid inflating
a bunch of GObjects just to have them immediate discarded.